+++ src/arrayTypes/index.js
@@ -1,0 +1,6 @@
+
+export default {
+  Boolean: "__honeytypes.arrayTypes.boolean",
+  Number: "__honeytypes.arrayTypes.number",
+  String: "__honeytypes.arrayTypes.string"
+}