+++ src/utils/nap/index.js
@@ -1,9 +1,9 @@
 
-const wait = (duration = 0) => {
+const nap = (duration = 0) => {
 
 	if (!isDurationUsable(duration)) {
 		duration = 0;
-		console.warn(`Redux-Honey: \n unable to use method wait(). Passed duration needs to be a number.`);
+		console.warn(`Redux-Honey: \n unable to use method nap(). Passed duration needs to be a number.`);
 	}
 
 	return new Promise(resolve => setTimeout(resolve, duration))
@@ -13,4 +13,4 @@ const isDurationUsable = duration => (
 	!isNaN(duration)
 )
 
-export default wait;
+export default nap;